Personal Plane-Chasing Project
I couldn't upload the full project because GitHub only allows up to 25mb and the actual file size even when (compressed) is around 120mb so, The PlaneChasing.csproj is the code I have written in only a days work that runs the project when fully put together. I have also include the PlaneChasing.sln and Assembly files.